Tribes: Some refactors
Splitting up the monolithic code into separate files and functions.
